% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodonAllFile.R
\name{CodonAllFile}
\alias{CodonAllFile}
\title{CodonAllFile}
\usage{
CodonAllFile(Adj = NULL, out_dir_base = ".")
}
\arguments{
\item{Adj}{Data frame containing adjusted mutation percentages.
Defaults to the global `Adj` object if not supplied.}

\item{out_dir_base}{Base analysis directory. Defaults to the current working
directory (`"."`). Output files are written to the `Overview/CodonFile`
subdirectory within this directory.}
}
\value{
creates csv file for each sample
}
\description{
File that can be used for further processing
}
