% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AdjCreateBarChart.R
\name{AdjCreateBarChart}
\alias{AdjCreateBarChart}
\title{AdjCreateBarChart}
\arguments{
\item{"colour"}{Color used to fill the bars. Default is `"skyblue"`.}
}
\value{
No return value. A bar chart PNG is saved to `./Plots/AdjustedPlots/Adjusted_Barchart_offTarget.png`.
}
\description{
Creates a bar chart of adjusted off-target percentages per sample
}
\examples{
\dontrun{
AdjCreateBarChart(colour = "skyblue")
}
}
