% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AdjOffTarget.R
\name{AdjOffTarget}
\alias{AdjOffTarget}
\title{AdjOffTarget}
\value{
No return value. Saves a CSV summary file in `./Overview/Adjusted`.
}
\description{
Calculates mean adjusted off-target heteroplasmy percentages per sample,
and merges with corresponding on-target percentages and average coverage.
}
\examples{
\dontrun{
AdjOffTarget()
}
}
\keyword{adjusted,}
\keyword{heteroplasmy,}
\keyword{off-target,}
\keyword{summary}
