% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AdjScatter.R
\name{AdjScatter}
\alias{AdjScatter}
\title{AdjScatter}
\usage{
AdjScatter(min_threshold = 0, labelPercentage = 10)
}
\arguments{
\item{min_threshold}{Numeric; minimum adjusted heteroplasmy percentage to display (default = 0).}

\item{labelPercentage}{Numeric; threshold above which positions are labeled on the plot (default = 10).}
}
\description{
Creates scatterplots showing adjusted heteroplasmy levels for each sample,
highlighting and labeling positions above a user-defined threshold.
}
\examples{
\dontrun{
AdjScatter(min_threshold = 0, labelPercentage = 10)
}
}
\keyword{heteroplasmy,}
\keyword{labeling}
\keyword{plot,}
\keyword{scatter,}
