% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bystander.R
\name{Bystander}
\alias{Bystander}
\title{Bystander}
\usage{
Bystander(
  BystanderDistance = 10,
  title = "Bystander effect",
  xlab = "mtDNA position",
  ylab = " "
)
}
\arguments{
\item{BystanderDistance}{Integer. Number of positions upstream/downstream from the on-target site to include.}

\item{title}{Character. Title of the heatmap plot. Default is "Bystander effect".}

\item{xlab}{Character. Label for the x-axis. Default is "mtDNA position".}

\item{ylab}{Character. Label for the y-axis. Default is "" (blank).}
}
\value{
No return value. Saves a heatmap plot to `./Plots/HeatmapBystanderEffect.png`.
}
\description{
Creates a heatmap of adjusted heteroplasmy percentages for positions surrounding the on-target site.
Highlights potential bystander effects within a user-defined window.
}
\examples{
\dontrun{
Bystander(BystanderDistance = 10)
}
}
