% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bystander_only_relevant.R
\name{bystander_only_relevant}
\alias{bystander_only_relevant}
\title{bystander_only_relevant}
\usage{
bystander_only_relevant(
  BystanderDistance = 10,
  title = "Bystander effect",
  xlab = "mtDNA position",
  ylab = " ",
  fill_colours = c("white", "lightblue", "darkblue"),
  fill_values  = c(0, 30, 100)
)
}
\arguments{
\item{BystanderDistance}{Integer. Number of positions upstream/downstream from the on-target site to include.}

\item{title}{Character. Title of the heatmap plot. Default is "Bystander effect".}

\item{xlab}{Character. Label for the x-axis. Default is "mtDNA position".}

\item{ylab}{Character. Label for the y-axis. Default is "" (blank).}
}
\value{
No return value. Saves a heatmap plot to `./Plots/HeatmapBystanderEffect_only_relevant.png`.
}
\description{
Creates a heatmap of adjusted heteroplasmy percentages for potential target sites surrounding the on-target site.
Highlights potential bystander effects within a user-defined window.
}
\examples{
\dontrun{
bystander_only_relevant(BystanderDistance = 10)
}
}
