% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalcAllMutations.R
\name{CalcAllMutations}
\alias{CalcAllMutations}
\title{CalcAllMutations}
\usage{
CalcAllMutations(AllReportFiles)
}
\arguments{
\item{AllReportFiles}{list with csv files}
}
\value{
Describe what the function returns
}
\description{
Calculate all mutations
}
