% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_create_dir.R
\name{check_create_dir}
\alias{check_create_dir}
\title{check_create_dir}
\usage{
check_create_dir(the_dir = "./subfolder")
}
\arguments{
\item{the_dir}{directory. Default ./subfolder}
}
\description{
This function allows you create and name a subdirectory
}
\examples{
check_create_dir("./name_subdir")
}
\keyword{(sub)directory}
\keyword{create}
