% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodonFile.R
\name{CodonFile}
\alias{CodonFile}
\title{CodonFile}
\usage{
CodonFile()
}
\description{
This function allows you calculate off target effects after removal of presumable polymorphisms
}
\examples{
\dontrun{
# Usage example
CodonFile()
}
}
