% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoverageCalc.R
\name{CoverageCalc}
\alias{CoverageCalc}
\title{CoverageCalc}
\usage{
CoverageCalc(CoverageFiles)
}
\arguments{
\item{CoverageFiles}{A character vector of file paths to coverage summary files (CSV).}
}
\value{
None. Used for its side effect: writes 'Coverage.csv' to disk.
}
\description{
Combines coverage data from multiple files into a single overview file.
Strips unnecessary parts from filenames, renames columns, and saves the result as "Coverage.csv".
}
\examples{
\dontrun{
CoverageFiles <- c("./data/coverage_file1.csv", "./data/coverage_file2.csv")
CoverageCalc(CoverageFiles)
}
}
\keyword{combine}
\keyword{coverage,}
\keyword{directory,}
