% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DdCBE_df_All.R
\name{DdCBE_df_All}
\alias{DdCBE_df_All}
\title{DdCBE_df_All}
\usage{
DdCBE_df_All(min_threshold = 0, max_threshold = 60, controls = 2)
}
\arguments{
\item{min_threshold}{Numeric; minimum heteroplasmy level to take into account, Default = 0}

\item{max_threshold}{Numeric; maximum heteroplasmy level in the controls to take into account, Default = 60}

\item{controls}{Numeric; number of controls that need to have heteroplasmy level above max_threshold. Default = 2}
}
\value{
Describe what the function returns
}
\description{
What the function does.
}
