% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DepthFile.R
\name{DepthFile}
\alias{DepthFile}
\title{DepthFile}
\usage{
DepthFile(fromP = 1, toP = NULL)
}
\arguments{
\item{fromP}{Integer, starting position for the range of interest. Default is 1.}

\item{toP}{Integer, ending position for the range of interest. If not supplied, the max position found is used.}
}
\description{
Creates an overview file containing the depth and heteroplasmy percentage per position
for each sample in the specified position range.
}
\examples{
\dontrun{
DepthFile(fromP = 300, toP = 500)
}
}
\keyword{depth,}
\keyword{heteroplasmy,}
\keyword{overview}
