% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FilterATEdits.R
\name{FilterATEdits}
\alias{FilterATEdits}
\title{FilterATEdits}
\usage{
FilterATEdits(csvFiles, out_dir = "./percentages")
}
\arguments{
\item{csvFiles}{list with csv files}

\item{out_dir}{output directory. Default "./percentages"}
}
\value{
Describe what the function returns
}
\description{
#' Filter for adenine base editing
}
