% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FilterCGEdits.R
\name{FilterCGEdits}
\alias{FilterCGEdits}
\title{FilterCGEdits}
\usage{
FilterCGEdits(csvFiles, OntargetPosition, out_dir_base = ".")
}
\arguments{
\item{csvFiles}{list with csv files}

\item{OntargetPosition}{Numeric; position within the genome}

\item{out_dir_base}{output directory. Default "./percentages"}
}
\value{
Describe what the function returns
}
\description{
Filter for CG editing
}
