% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/off_target_count_summary.R
\name{off_target_count_summary}
\alias{off_target_count_summary}
\title{off_target_count_summary}
\usage{
off_target_count_summary(
  Adj = get("Adj", envir = .GlobalEnv),
  SampleList = get("SampleList", envir = .GlobalEnv),
  lower = 0.1,
  upper = 70,
  condition_levels = c("control", "treated"),
  out_dir = ".",
  csv_prefix = "MutationCount",
  plot_file = "MutationCount_grouped.pdf",
  make_plot = TRUE,
  combined_col_in_samplelist = "CombinedName",
  condition_col = "Condition"
)
}
\arguments{
\item{lower}{minimum heteroplasmy level to take into account}

\item{upper}{maximum heteroplasmy level to take into account}

\item{condition_levels}{Default control and treated}

\item{out_dir}{output directory}

\item{csv_prefix}{Default "MutationCount"}

\item{plot_file}{pdf name. Default "MutationCount_grouped.pdf"}
}
\value{
Describe what the function returns
}
\description{
What the function does.
}
