% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OnTargetCalcABE.R
\name{OnTargetCalcABE}
\alias{OnTargetCalcABE}
\title{OnTargetCalcABE}
\usage{
OnTargetCalcABE(MinXFiles)
}
\arguments{
\item{MinXFiles}{A character vector of file paths}
}
\description{
This function allows you to perform all necessary calculations for adenine base editing on and off target effects
}
\examples{
\dontrun{
# Usage example
file_paths <- c("file1.csv", "file2.csv", "file3.csv")
OnTargetCalcAB(file_paths)
}
}
\keyword{(default}
\keyword{30)}
\keyword{=}
\keyword{Load}
\keyword{and}
\keyword{counts}
\keyword{depth}
\keyword{filter}
\keyword{for}
\keyword{minimum}
