% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_OffTarget_histograms.R
\name{plot_OffTarget_histograms}
\alias{plot_OffTarget_histograms}
\title{plot_offTarget_histograms}
\usage{
plot_OffTarget_histograms(
  bw = 1,
  min_pct = 0,
  max_pct = Inf,
  Adj = NULL,
  OntargetPosition = NULL
)
}
\arguments{
\item{bw}{Bin width}

\item{min_pct}{Minimum heteroplasmy percentage}

\item{max_pct}{Maximum heteroplasmy percentage}

\item{Adj}{Data frame containing editing percentages.
Defaults to the global `Adj` object if not supplied.}

\item{OntargetPosition}{Numeric. Genomic position of the intended on-target
editing site. Defaults to the global `OntargetPosition` object if not supplied.}
}
\value{
Describe what the function returns
}
\description{
Make histograms
}
