% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_OffTarget_histograms.R
\name{plot_OffTarget_histograms}
\alias{plot_OffTarget_histograms}
\title{plot_offTarget_histograms}
\usage{
plot_OffTarget_histograms(bw = 1, min_pct = 0, max_pct = Inf)
}
\arguments{
\item{bw}{Bin width}

\item{min_pct}{Minimum heteroplasmy percentage}

\item{max_pct}{Maximum heteroplasmy percentage}
}
\value{
Describe what the function returns
}
\description{
Make histograms
}
